refactor: 프로젝트 구조 변경
- appController -> appService -> domainService -> domainRepository
