--- src/main/java/nextstep/application/JwtTokenProvider.java
@@ -1,64 +1,0 @@
-package nextstep.application;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import java.util.Date;
-import java.util.List;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-@Component
-public class JwtTokenProvider {
-
-    @Value("${security.jwt.token.secret-key}")
-    private String secretKey;
-
-    @Value("${security.jwt.token.expire-length}")
-    private long validityInMilliseconds;
-
-    public String createToken(String principal, List<String> roles) {
-        Claims claims = Jwts.claims().setSubject(principal);
-        Date now = new Date();
-        Date validity = new Date(now.getTime() + validityInMilliseconds);
-
-        return Jwts.builder()
-            .setClaims(claims)
-            .setIssuedAt(now)
-            .setExpiration(validity)
-            .claim("roles", roles)
-            .signWith(SignatureAlgorithm.HS256, secretKey)
-            .compact();
-    }
-
-    public String getPrincipal(String token) {
-        return Jwts.parser()
-            .setSigningKey(secretKey)
-            .parseClaimsJws(token)
-            .getBody()
-            .getSubject();
-    }
-
-    public List<String> getRoles(String token) {
-        return Jwts.parser()
-            .setSigningKey(secretKey)
-            .parseClaimsJws(token)
-            .getBody()
-            .get("roles", List.class);
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jws<Claims> claims = Jwts.parser()
-                .setSigningKey(secretKey)
-                .parseClaimsJws(token);
-
-            return !claims.getBody().getExpiration().before(new Date());
-
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
-    }
-}
