--- src/main/java/nextstep/application/ScheduleService.java
@@ -1,102 +1,0 @@
-package nextstep.application;
-
-import java.util.List;
-import java.util.stream.Collectors;
-import nextstep.domain.Schedule;
-import nextstep.domain.repository.ScheduleRepository;
-import nextstep.exception.ScheduleException;
-import nextstep.presentation.dto.reservation.ReservationRequest;
-import nextstep.presentation.dto.reservation.ReservationResponse;
-import nextstep.presentation.dto.schedule.ScheduleRequest;
-import nextstep.presentation.dto.schedule.ScheduleResponse;
-import nextstep.presentation.dto.theme.ThemeResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Transactional(readOnly = true)
-@Service
-public class ScheduleService {
-
-    private final ThemeService themeService;
-    private final ReservationService reservationService;
-    private final ScheduleRepository scheduleRepository;
-
-    public ScheduleService(
-        ThemeService themeService,
-        ReservationService reservationService,
-        ScheduleRepository scheduleRepository
-    ) {
-        this.themeService = themeService;
-        this.reservationService = reservationService;
-        this.scheduleRepository = scheduleRepository;
-    }
-
-    @Transactional
-    public Long make(ScheduleRequest request) {
-        Long reservationId = reservationService.make(toReservationRequest(request));
-
-        scheduleRepository.save(toSchedule(request, reservationId));
-        return scheduleRepository.findByReservationId(reservationId)
-            .map(Schedule::getId)
-            .orElseThrow(() -> new ScheduleException("존재하지 않는 스케줄입니다."));
-    }
-
-    private ReservationRequest toReservationRequest(ScheduleRequest request) {
-        return new ReservationRequest(request.getDate(), request.getTime());
-    }
-
-    private Schedule toSchedule(ScheduleRequest request, Long reservationId) {
-        return new Schedule(request.getThemeId(), reservationId);
-    }
-
-    public List<ScheduleResponse> checkAll(Long themeId, String date) {
-        List<ReservationResponse> reservations = reservationService.checkAll(date);
-        ThemeResponse theme = themeService.checkBy(themeId);
-
-        return reservations.stream()
-            .map(reservation -> toScheduleResponse(themeId, reservation, theme))
-            .collect(Collectors.toList());
-    }
-
-    private ScheduleResponse toScheduleResponse(
-        Long themeId,
-        ReservationResponse reservation,
-        ThemeResponse theme
-    ) {
-        Schedule schedule = scheduleRepository.findBy(themeId, reservation.getId())
-            .orElseThrow(() -> new ScheduleException("존재하지 않는 스케줄입니다."));
-
-        return toScheduleResponse(theme, reservation, schedule);
-    }
-
-    private ScheduleResponse toScheduleResponse(
-        ThemeResponse theme,
-        ReservationResponse reservation,
-        Schedule schedule
-    ) {
-        return new ScheduleResponse(
-            schedule.getId(),
-            theme,
-            reservation.getDate(),
-            reservation.getTime()
-        );
-    }
-
-    @Transactional
-    public void cancel(Long id) {
-        scheduleRepository.findByScheduleId(id)
-            .ifPresentOrElse(schedule -> {
-                if (reservationService.exist(schedule.getReservationId())) {
-                    throw new ScheduleException("스케줄을 삭제할 수 없습니다.");
-                }
-                scheduleRepository.delete(id);
-            }, () -> {
-                throw new ScheduleException("존재하지 않는 스케줄입니다.");
-            });
-    }
-
-    @Transactional
-    public void cancelAll() {
-        scheduleRepository.deleteAll();
-    }
-}
