+++ src/main/java/nextstep/application/controller/MemberController.java
@@ -1,0 +1,35 @@
+package nextstep.application.controller;
+
+import java.net.URI;
+import nextstep.application.service.MemberService;
+import nextstep.application.dto.member.MemberRequest;
+import nextstep.application.dto.member.MemberResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/members")
+public class MemberController {
+
+    private final MemberService memberService;
+
+    public MemberController(MemberService memberService) {
+        this.memberService = memberService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> createMember(@RequestBody MemberRequest memberRequest) {
+        Long id = memberService.create(memberRequest);
+        return ResponseEntity.created(URI.create("/members/" + id)).build();
+    }
+
+    @GetMapping("/me")
+    public ResponseEntity<MemberResponse> me() {
+        MemberResponse response = memberService.findById(1L);
+        return ResponseEntity.ok(response);
+    }
+}
