+++ src/main/java/nextstep/application/dto/schedule/ScheduleRequest.java
@@ -1,0 +1,29 @@
+package nextstep.application.dto.schedule;
+
+public class ScheduleRequest {
+
+    private Long themeId;
+    private String date;
+    private String time;
+
+    private ScheduleRequest() {
+    }
+
+    public ScheduleRequest(Long themeId, String date, String time) {
+        this.themeId = themeId;
+        this.date = date;
+        this.time = time;
+    }
+
+    public Long getThemeId() {
+        return themeId;
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public String getTime() {
+        return time;
+    }
+}
