+++ src/main/java/nextstep/application/dto/schedule/ScheduleResponse.java
@@ -1,0 +1,37 @@
+package nextstep.application.dto.schedule;
+
+import nextstep.application.dto.theme.ThemeResponse;
+
+public class ScheduleResponse {
+
+    private Long id;
+    private ThemeResponse theme;
+    private String date;
+    private String time;
+
+    private ScheduleResponse() {
+    }
+
+    public ScheduleResponse(Long id, ThemeResponse theme, String date, String time) {
+        this.id = id;
+        this.theme = theme;
+        this.date = date;
+        this.time = time;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public ThemeResponse getTheme() {
+        return theme;
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public String getTime() {
+        return time;
+    }
+}
