+++ src/main/java/nextstep/application/dto/theme/ThemeResponse.java
@@ -1,0 +1,35 @@
+package nextstep.application.dto.theme;
+
+public class ThemeResponse {
+
+    private Long id;
+    private String name;
+    private String desc;
+    private Integer price;
+
+    private ThemeResponse() {
+    }
+
+    public ThemeResponse(Long id, String name, String desc, Integer price) {
+        this.id = id;
+        this.name = name;
+        this.desc = desc;
+        this.price = price;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDesc() {
+        return desc;
+    }
+
+    public Integer getPrice() {
+        return price;
+    }
+}
