+++ src/main/java/nextstep/common/exception/AuthenticationException.java
@@ -1,0 +1,4 @@
+package nextstep.common.exception;
+
+public class AuthenticationException extends RuntimeException {
+}
