+++ src/main/java/nextstep/common/exception/ReservationException.java
@@ -1,0 +1,8 @@
+package nextstep.common.exception;
+
+public class ReservationException extends RuntimeException {
+
+    public ReservationException(String message) {
+        super(message);
+    }
+}
