+++ src/main/java/nextstep/common/exception/ScheduleException.java
@@ -1,0 +1,8 @@
+package nextstep.common.exception;
+
+public class ScheduleException extends RuntimeException {
+
+    public ScheduleException(String message) {
+        super(message);
+    }
+}
