+++ src/main/java/nextstep/common/exception/ThemeException.java
@@ -1,0 +1,8 @@
+package nextstep.common.exception;
+
+public class ThemeException extends RuntimeException {
+
+    public ThemeException(String message) {
+        super(message);
+    }
+}
