+++ src/main/java/nextstep/domain/repository/ScheduleRepository.java
@@ -7,9 +7,9 @@ public interface ScheduleRepository {
 
     void save(Schedule schedule);
 
-    Optional<Schedule> findByReservationId(Long reservationId);
+    Optional<Schedule> findByReservation(Long id);
 
-    Optional<Schedule> findByScheduleId(Long id);
+    Optional<Schedule> findBySchedule(Long id);
 
     Optional<Schedule> findBy(Long themeId, Long reservationId);
