+++ src/main/java/nextstep/domain/service/MemberDomainService.java
@@ -1,0 +1,5 @@
+package nextstep.domain.service;
+
+public class MemberDomainService {
+
+}
