--- src/main/java/nextstep/exception/AuthenticationException.java
@@ -1,4 +1,0 @@
-package nextstep.exception;
-
-public class AuthenticationException extends RuntimeException {
-}
