--- src/main/java/nextstep/exception/ReservationException.java
@@ -1,8 +1,0 @@
-package nextstep.exception;
-
-public class ReservationException extends RuntimeException {
-
-    public ReservationException(String message) {
-        super(message);
-    }
-}
