--- src/main/java/nextstep/exception/ScheduleException.java
@@ -1,8 +1,0 @@
-package nextstep.exception;
-
-public class ScheduleException extends RuntimeException {
-
-    public ScheduleException(String message) {
-        super(message);
-    }
-}
