--- src/main/java/nextstep/exception/ThemeException.java
@@ -1,8 +1,0 @@
-package nextstep.exception;
-
-public class ThemeException extends RuntimeException {
-
-    public ThemeException(String message) {
-        super(message);
-    }
-}
