+++ src/test/java/nextstep/acceptance/MemberE2ETest.java
@@ -1,7 +1,7 @@
 package nextstep.acceptance;
 
 import io.restassured.RestAssured;
-import nextstep.presentation.dto.member.MemberRequest;
+import nextstep.application.dto.member.MemberRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
